+++ src/main/java/Help.java
@@ -43,7 +43,7 @@ public class Help {
         }
         return resultMat;
     }
-    public static int ManhattanDistance(Integer stateInt){
+    public static float ManhattanDistance(Integer stateInt){
         Integer[][] goal = {{1,1,0},{2,2,0},{3,0,1},{4,1,1},{5,2,1},{6,0,2},{7,1,2},{8,2,2}};
         int result = 0;
         int element = 0;
@@ -61,18 +61,19 @@ public class Help {
         }
         return result;
     }
-    public static int EuclideanDistance(Integer stateInt){
+
+    public static float EuclideanDistance(Integer stateInt){
         Integer[][] goal = {{1,1,0},{2,2,0},{3,0,1},{4,1,1},{5,2,1},{6,0,2},{7,1,2},{8,2,2}};
-        int result = 0;
+        float result = 0;
         int element = 0;
-        int euclidean = 0;
+        float euclidean = 0;
         Integer[][] resultMat = new Integer[3][3];
         resultMat = toMatrix(stateInt);
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 3; j++) {
                 element = resultMat[i][j];
                 if (element>0){
-                    euclidean = (int) Math.sqrt(Math.pow(goal[element-1][1]-j,2) + Math.pow(goal[element-1][2]-i,2));
+                    euclidean = (float) Math.sqrt(Math.pow(goal[element-1][1]-j,2) + Math.pow(goal[element-1][2]-i,2));
                     result = result + euclidean;
                 }
             }
